+++ react/components/admin/institutional/Form/Form.tsx
@@ -5,7 +5,6 @@ import { defineMessages, FormattedMessage, injectIntl } from 'react-intl'
 import { Button, Input, Textarea, IconEdit } from 'vtex.styleguide'
 
 import { RouteContentFromData } from './index'
-import { slugify } from './utils'
 
 import FormFieldSeparator from '../../FormFieldSeparator'
 import { FormErrors } from '../../pages/Form/typings'
@@ -13,6 +12,7 @@ import SeparatorWithLine from '../../pages/SeparatorWithLine'
 
 import RichTextEditor from '../../../RichTextEditor/index'
 import SeoPreview from '../../../SeoPreview'
+import { slugify } from '../../../helpers'
 
 interface CustomProps {
   data: RouteFormData & RouteContentFromData
